+++ modules/canvaskit/canvaskit_bindings.cpp
@@ -723,12 +723,12 @@ void deleteJSTexture(SkImage::ReleaseContext rc) {
 
 class WebGLTextureImageGenerator : public SkImageGenerator {
 public:
-    WebGLTextureImageGenerator(int width, int height, JSObject callbackObj):
-            SkImageGenerator(SkImageInfo::MakeN32Premul(width, height)),
+    WebGLTextureImageGenerator(SkImageInfo ii, JSObject callbackObj):
+            SkImageGenerator(ii),
             fCallback(callbackObj) {}
 
     ~WebGLTextureImageGenerator() {
-        // This cleans up the associated TextureSource at is used to make the texture
+        // This cleans up the associated TextureSource that is used to make the texture
         // (i.e. "makeTexture" below). We expect this destructor to be called when the
         // SkImage that this Generator belongs to is destroyed.
         fCallback.call<void>("freeSrc");
@@ -785,8 +785,8 @@ private:
 // callbackObj has two functions in it, one to create a texture "makeTexture" and one to clean up
 // the underlying texture source "freeSrc". This way, we can create WebGL textures for each
 // surface/WebGLContext that the image is used on (we cannot share WebGLTextures across contexts).
-sk_sp<SkImage> MakeImageFromGenerator(int width, int height, JSObject callbackObj) {
-    auto gen = std::make_unique<WebGLTextureImageGenerator>(width, height, callbackObj);
+sk_sp<SkImage> MakeImageFromGenerator(SimpleImageInfo ii, JSObject callbackObj) {
+    auto gen = std::make_unique<WebGLTextureImageGenerator>(toSkImageInfo(ii), callbackObj);
     return SkImage::MakeFromGenerator(std::move(gen));
 }
 #endif // SK_GL